Reorganize user-home files
