+++ client/components/index.js
@@ -4,7 +4,7 @@
  * having to figure out which file they belong to!
  */
 export {default as Navbar} from './navbar'
-export {default as UserHome} from './user-home'
+export {default as UserHome} from './user-home/user-home'
 export {Login, Signup} from './auth-form'
 export {default as PlayerAddForm} from './player-add-form'
 export {default as PlayerList} from './player-list'
