+++ client/components/user-home/user-home.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
-import {getEventsByStatus, fetchAllEvents} from '../store/event'
-import {EventCard, EventActionButton} from '../components'
-import InstructionPanel from './user-home/instructionsPanel'
+import {getEventsByStatus, fetchAllEvents} from '../../store/event'
+import {EventCard, EventActionButton} from '..'
+import InstructionPanel from './instructionsPanel'
 
 /**
  * COMPONENT
